Rename the module to 'successors'
