+++ src/stream/successors.rs
